requirements.txt
fastapi
uvicorn
supabase==2.23.0
python-dotenv
pydantic
requests
stripe
openai
python-whois
mcp[fastmcp]
fastmcp>=2.12.0
reportlab
redis
python-multipart
validators
loguru
pytz
gunicorn
PyJWT
sentry-sdk
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-redis
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http
logtail-python>=0.3.0
slowapi>=0.1.9
groq
boto3
cryptography
networkx
rfc3161ng